added new suggestions
