+++ larpmanager/locale/cs/LC_MESSAGES/django.po
@@ -1272,6 +1272,9 @@ msgstr "speedlarp"
 msgid "surcharge"
 msgstr "příplatek"
 
+msgid "this link"
+msgstr "tento odkaz"
+
 msgid "Accounting"
 msgstr "Účetnictví"
 
@@ -2744,9 +2747,6 @@ msgstr "Další vstupenky"
 msgid "Approved character"
 msgstr "Schválená postava"
 
-msgid "Can't find a user?"
-msgstr "Nemůžete najít uživatele?"
-
 msgid "Character portrait"
 msgstr "Portrét postavy"
 
@@ -3153,9 +3153,6 @@ msgstr "Historické účetnictví"
 
 msgid "If absent, indicate 0"
 msgstr "Pokud chybí, uveďte 0"
-
-msgid "Links to be accessed:"
-msgstr "Odkazy, ke kterým je třeba přistupovat:"
 
 msgid "List of other users: "
 msgstr "Seznam dalších uživatelů:"
@@ -4711,6 +4708,9 @@ msgstr "Pokud je zaškrtnuto: zobrazuje slogan pro každou událost."
 msgid "If checked: shows the website for each event"
 msgstr "Pokud je zaškrtnuto: zobrazí webové stránky pro každou událost"
 
+msgid "If you don't see an user, ask them to access"
+msgstr "Pokud uživatele nevidíte, požádejte ho o přístup k."
+
 msgid "Manage invoices of payments and approve them"
 msgstr "Správa faktur a jejich schvalování."
 
@@ -6211,6 +6211,11 @@ msgid ""
 msgstr "Tato stránka umožňuje konfigurovat funkce."
 
 msgid ""
+"To assign a user to a role, they must first be added to the organization."
+msgstr ""
+"Chcete-li uživateli přiřadit roli, musí být nejprve přidán do organizace."
+
+msgid ""
 "Total amount of payment to be received by this date (0 = all outstanding)"
 msgstr ""
 "Celková částka platby, kterou je třeba obdržet do tohoto data (0 = všechny "
@@ -6945,13 +6950,6 @@ msgid ""
 msgstr "Nahrajte podepsanou žádost a fotografii dokladu totožnosti."
 
 msgid ""
-"this figure is the net that should result as the availability of financial "
-"instruments"
-msgstr ""
-"tento údaj je čistá hodnota, která by měla být výsledkem dostupnosti "
-"finančních nástrojů."
-
-msgid ""
 "Adds a page displaying member mailing lists, sorted by members' newsletter "
 "preferences."
 msgstr ""
@@ -7482,6 +7480,13 @@ msgid ""
 msgstr ""
 "Nepovinné: nahrajte svou fotografii spojenou s vaší postavou speciálně pro "
 "tuto událost!"
+
+msgid ""
+"This is the overall amount that should be held in the organization’s bank "
+"and financial accounts"
+msgstr ""
+"Jedná se o celkovou částku, která by měla být uložena na bankovních a "
+"finančních účtech organizace."
 
 msgid ""
 "Unique code for internal use - not visible. Indicate a string without spaces"
@@ -8297,13 +8302,6 @@ msgstr ""
 " si vyberou tento tip."
 
 msgid ""
-"this figure is the associative positive, the operating margin. It takes into"
-" account the budgets of the completed runs"
-msgstr ""
-"tento údaj je přidružené kladné, provozní rozpětí. Zohledňuje rozpočty "
-"dokončených běhů."
-
-msgid ""
 "Are you a doctor, a nurse, or a licensed rescuer? We can ask you to "
 "intervene in case accidents occur during the event?"
 msgstr ""
@@ -8424,6 +8422,13 @@ msgid ""
 msgstr ""
 "Na základě zadaných hodnot se simuluje optimální alokace; před finálním "
 "načtením ji můžete spustit libovolně mnohokrát."
+
+msgid ""
+"To define which data will be asked in the profile form to the users once "
+"they sign up, access the profile management page:"
+msgstr ""
+"Chcete-li definovat, které údaje se budou uživatelům po registraci "
+"zobrazovat v profilovém formuláři, přejděte na stránku správy profilu:"
 
 msgid ""
 "Lets you specify a set of users who receive payment confirmation emails for "
@@ -8571,6 +8576,13 @@ msgid ""
 msgstr ""
 "Tato stránka shrnuje sbírky otevřené hráči. Můžete je aktualizovat uvedením "
 "příjemce a vyplatit peníze."
+
+msgid ""
+"This is the associative positive operating margin, taking into account the "
+"budgets of completed runs, minus credits or tokens issued"
+msgstr ""
+"Jedná se o přidruženou kladnou provozní marži, která zohledňuje rozpočty "
+"dokončených běhů po odečtení kreditů nebo vydaných žetonů."
 
 msgid ""
 "Adds a quiz-style system for workshop questions. Organizers can define "
@@ -8745,6 +8757,14 @@ msgstr ""
 "právě kontrolují pracovníci. V tuto chvíli nemusíte nic dělat!"
 
 msgid ""
+"To define the registration form, and set up any number of registration "
+"questions and their options, access the registration form management page:"
+msgstr ""
+"Chcete-li definovat registrační formulář a nastavit libovolný počet "
+"registračních otázek a jejich možností, přejděte na stránku pro správu "
+"registračního formuláře:"
+
+msgid ""
 "If you press once, it will be *included* in the search. If a character "
 "belongs to at least one inclusion category, it will be shown in the search."
 msgstr ""
@@ -8782,14 +8802,6 @@ msgstr ""
 "Platbu můžete provést během několika minut <a href= %(url)s'>zde</a>. Dejte "
 "nám vědět, pokud narazíte na nějaké problémy nebo pokud vám můžeme jakkoli "
 "pomoci!"
-
-msgid ""
-"According to LarpManager's privacy rules, you can only view users who have "
-"granted LarpManager permission to share their data with your organisation."
-msgstr ""
-"Podle pravidel ochrany osobních údajů LarpManageru můžete zobrazit pouze "
-"uživatele, kteří udělili LarpManageru povolení sdílet jejich údaje s vaší "
-"organizací."
 
 msgid ""
 "At the top, you'll find shortcuts to access your organization dashboard "
@@ -8869,6 +8881,15 @@ msgstr ""
 "patří byť jen do jedné vylučovací kategorie, nebude ve vyhledávání zobrazen."
 
 msgid ""
+"To set up the gateway payment available to players, to let them pay the "
+"registration fee through the platform, access the payment settings "
+"management page:"
+msgstr ""
+"Chcete-li nastavit platbu přes platební bránu, která je hráčům k dispozici, "
+"aby mohli zaplatit registrační poplatek prostřednictvím platformy, přejděte "
+"na stránku pro správu nastavení plateb:"
+
+msgid ""
 "Indicates how you can be recognized, if you will be found near some point "
 "specific, if you have a lot of luggage: any information that might help us "
 "help you"
@@ -8892,15 +8913,6 @@ msgstr ""
 "Na této stránce jsou shrnuty dotazy uživatelů rozdělené podle událostí, "
 "kterých se týkají. Na otázku můžete odpovědět, nebo ji zavřít, pokud není "
 "třeba provádět žádné další kroky."
-
-msgid ""
-"Ask your co-workers to access this link; after their confirmation, their "
-"data will be shared with your organisation, and you will be able to add them"
-" to roles."
-msgstr ""
-"Požádejte své spolupracovníky o přístup k tomuto odkazu; po jejich potvrzení"
-" budou jejich údaje sdíleny s vaší organizací a vy je budete moci přidat do "
-"rolí."
 
 #, python-format
 msgid ""
