+++ larpmanager/locale/de/LC_MESSAGES/django.po
@@ -1272,6 +1272,9 @@ msgstr "speedlarp"
 msgid "surcharge"
 msgstr "Aufpreis"
 
+msgid "this link"
+msgstr "dieser Link"
+
 msgid "Accounting"
 msgstr "Buchhaltung"
 
@@ -2744,9 +2747,6 @@ msgstr "Zusätzliche Tickets"
 msgid "Approved character"
 msgstr "Zulässiges Zeichen"
 
-msgid "Can't find a user?"
-msgstr "Sie können einen Benutzer nicht finden?"
-
 msgid "Character portrait"
 msgstr "Charakter-Porträt"
 
@@ -3153,9 +3153,6 @@ msgstr "Historische Buchführung"
 
 msgid "If absent, indicate 0"
 msgstr "Falls nicht vorhanden, 0 angeben"
-
-msgid "Links to be accessed:"
-msgstr "Links, die aufgerufen werden können:"
 
 msgid "List of other users: "
 msgstr "Liste der anderen Benutzer:"
@@ -4719,6 +4716,9 @@ msgstr "Wenn markiert: zeigt die Tagline für jedes Ereignis an"
 
 msgid "If checked: shows the website for each event"
 msgstr "Wenn angekreuzt: zeigt die Website für jedes Ereignis"
+
+msgid "If you don't see an user, ask them to access"
+msgstr "Wenn Sie keinen Benutzer sehen, fragen Sie ihn nach dem Zugang"
 
 msgid "Manage invoices of payments and approve them"
 msgstr "Verwaltung und Genehmigung von Rechnungen und Zahlungen"
@@ -6538,6 +6538,12 @@ msgid ""
 msgstr "Auf dieser Seite können Sie Funktionen konfigurieren."
 
 msgid ""
+"To assign a user to a role, they must first be added to the organization."
+msgstr ""
+"Um einem Benutzer eine Rolle zuzuweisen, muss er zunächst zur Organisation "
+"hinzugefügt werden."
+
+msgid ""
 "Total amount of payment to be received by this date (0 = all outstanding)"
 msgstr ""
 "Gesamtbetrag der bis zu diesem Datum zu leistenden Zahlungen (0 = alles "
@@ -7340,13 +7346,6 @@ msgstr ""
 "Laden Sie Ihren unterschriebenen Antrag und ein Foto Ihres Ausweises hoch"
 
 msgid ""
-"this figure is the net that should result as the availability of financial "
-"instruments"
-msgstr ""
-"diese Zahl ist das Nettoergebnis, das sich aus der Verfügbarkeit von "
-"Finanzinstrumenten ergeben sollte"
-
-msgid ""
 "Adds a page displaying member mailing lists, sorted by members' newsletter "
 "preferences."
 msgstr ""
@@ -7899,6 +7898,13 @@ msgid ""
 msgstr ""
 "Optional: Laden Sie ein Foto von sich selbst hoch, das mit Ihrem Charakter "
 "speziell für dieses Ereignis verbunden ist!"
+
+msgid ""
+"This is the overall amount that should be held in the organization’s bank "
+"and financial accounts"
+msgstr ""
+"Dies ist der Gesamtbetrag, der auf den Bank- und Finanzkonten der "
+"Organisation gehalten werden sollte."
 
 msgid ""
 "Unique code for internal use - not visible. Indicate a string without spaces"
@@ -8729,13 +8735,6 @@ msgstr ""
 "Spielern angezeigt, die dieses Ticket ausgewählt haben."
 
 msgid ""
-"this figure is the associative positive, the operating margin. It takes into"
-" account the budgets of the completed runs"
-msgstr ""
-"Diese Zahl ist das assoziative Positiv, die operative Marge. Sie "
-"berücksichtigt die Budgets der abgeschlossenen Läufe"
-
-msgid ""
 "Are you a doctor, a nurse, or a licensed rescuer? We can ask you to "
 "intervene in case accidents occur during the event?"
 msgstr ""
@@ -8862,6 +8861,14 @@ msgid ""
 msgstr ""
 "Die optimale Zuteilung auf der Grundlage der eingestellten Werte wird "
 "simuliert; Sie können sie vor dem endgültigen Laden beliebig oft ausführen"
+
+msgid ""
+"To define which data will be asked in the profile form to the users once "
+"they sign up, access the profile management page:"
+msgstr ""
+"Um festzulegen, welche Daten im Profilformular von den Nutzern abgefragt "
+"werden, sobald sie sich angemeldet haben, rufen Sie die "
+"Profilverwaltungsseite auf:"
 
 msgid ""
 "Lets you specify a set of users who receive payment confirmation emails for "
@@ -9024,6 +9031,14 @@ msgstr ""
 "Auf dieser Seite werden die von Spielern eröffneten Sammlungen "
 "zusammengefasst. Sie können sie aktualisieren, indem Sie den Empfänger "
 "angeben und das Geld auszahlen."
+
+msgid ""
+"This is the associative positive operating margin, taking into account the "
+"budgets of completed runs, minus credits or tokens issued"
+msgstr ""
+"Dies ist die assoziative positive Betriebsmarge unter Berücksichtigung der "
+"Budgets der abgeschlossenen Läufe, abzüglich der ausgegebenen Kredite oder "
+"Token"
 
 msgid ""
 "Adds a quiz-style system for workshop questions. Organizers can define "
@@ -9217,6 +9232,14 @@ msgstr ""
 "zu diesem Zeitpunkt noch nichts tun!"
 
 msgid ""
+"To define the registration form, and set up any number of registration "
+"questions and their options, access the registration form management page:"
+msgstr ""
+"Um das Registrierungsformular zu definieren und eine beliebige Anzahl von "
+"Registrierungsfragen und deren Optionen einzurichten, rufen Sie die Seite "
+"zur Verwaltung des Registrierungsformulars auf:"
+
+msgid ""
 "If you press once, it will be *included* in the search. If a character "
 "belongs to at least one inclusion category, it will be shown in the search."
 msgstr ""
@@ -9257,14 +9280,6 @@ msgstr ""
 "Sie können die Zahlung in nur wenigen Minuten <a href= %(url)s'>hier</a> "
 "vornehmen. Lassen Sie uns wissen, wenn Sie Probleme haben oder wir Ihnen "
 "helfen können!"
-
-msgid ""
-"According to LarpManager's privacy rules, you can only view users who have "
-"granted LarpManager permission to share their data with your organisation."
-msgstr ""
-"Gemäß den Datenschutzbestimmungen von LarpManager können Sie nur Benutzer "
-"anzeigen, die LarpManager die Erlaubnis erteilt haben, ihre Daten mit Ihrer "
-"Organisation zu teilen."
 
 msgid ""
 "At the top, you'll find shortcuts to access your organization dashboard "
@@ -9352,6 +9367,15 @@ msgstr ""
 "der Suche nicht angezeigt."
 
 msgid ""
+"To set up the gateway payment available to players, to let them pay the "
+"registration fee through the platform, access the payment settings "
+"management page:"
+msgstr ""
+"Um die Gateway-Zahlung einzurichten, die den Spielern zur Verfügung steht, "
+"damit sie die Anmeldegebühr über die Plattform bezahlen können, gehen Sie "
+"auf die Verwaltungsseite für die Zahlungseinstellungen:"
+
+msgid ""
 "Indicates how you can be recognized, if you will be found near some point "
 "specific, if you have a lot of luggage: any information that might help us "
 "help you"
@@ -9378,15 +9402,6 @@ msgstr ""
 "Ereignissen, auf die sie sich beziehen, zusammengefasst. Sie können die "
 "Frage beantworten oder sie schließen, wenn keine weiteren Maßnahmen "
 "erforderlich sind."
-
-msgid ""
-"Ask your co-workers to access this link; after their confirmation, their "
-"data will be shared with your organisation, and you will be able to add them"
-" to roles."
-msgstr ""
-"Bitten Sie Ihre Mitarbeiter, auf diesen Link zuzugreifen; nach ihrer "
-"Bestätigung werden ihre Daten mit Ihrer Organisation geteilt, und Sie können"
-" sie zu Rollen hinzufügen."
 
 #, python-format
 msgid ""
