+++ larpmanager/locale/en/LC_MESSAGES/django.po
@@ -4397,8 +4397,8 @@ msgid ""
 msgstr ""
 
 msgid ""
-"this figure is the associative positive, the operating margin. It takes into "
-"account the budgets of the completed runs"
+"This is the associative positive operating margin, taking into account the "
+"budgets of completed runs, minus credits or tokens issued"
 msgstr ""
 
 msgid "Membership fees"
@@ -4414,8 +4414,8 @@ msgid "issued"
 msgstr ""
 
 msgid ""
-"this figure is the net that should result as the availability of financial "
-"instruments"
+"This is the overall amount that should be held in the organization’s bank "
+"and financial accounts"
 msgstr ""
 
 msgid "Transaction totals"
@@ -4674,10 +4674,14 @@ msgid ""
 "concerning the organisation."
 msgstr ""
 
-msgid "Can't find a user?"
+msgid ""
+"To assign a user to a role, they must first be added to the organization."
 msgstr ""
 
-msgid "Links to be accessed:"
+msgid "If you don't see an user, ask them to access"
+msgstr ""
+
+msgid "this link"
 msgstr ""
 
 msgid "Users"
@@ -6378,17 +6382,6 @@ msgid ""
 "concerning the event."
 msgstr ""
 
-msgid ""
-"According to LarpManager's privacy rules, you can only view users who have "
-"granted LarpManager permission to share their data with your organisation."
-msgstr ""
-
-msgid ""
-"Ask your co-workers to access this link; after their confirmation, their "
-"data will be shared with your organisation, and you will be able to add them "
-"to roles."
-msgstr ""
-
 msgid "Secret registration link"
 msgstr ""
 
@@ -7946,41 +7939,6 @@ msgstr ""
 msgid "Registration completed successfully!"
 msgstr ""
 
-msgid "Registrations on external link"
-msgstr ""
-
-msgid "Pre-registration active"
-msgstr ""
-
-msgid "Registrations opening not set"
-msgstr ""
-
-#, python-format
-msgid "Registrations opening at: %(date)s"
-msgstr ""
-
-msgid "Registrations open"
-msgstr ""
-
-msgid "Filler registrations"
-msgstr ""
-
-msgid "Waiting list registrations"
-msgstr ""
-
-msgid "Registration closed"
-msgstr ""
-
-msgid ""
-"There are no active events, to create a new one access the events management "
-"page:"
-msgstr ""
-
-msgid ""
-"To set the tickets that users can select during registration, access the "
-"tickets management page:"
-msgstr ""
-
 msgid "Request approved"
 msgstr ""
 
@@ -8034,6 +7992,57 @@ msgstr ""
 
 #, python-format
 msgid "Name: %(name)s, slug: %(slug)s, creator: %(user)s %(email)s"
+msgstr ""
+
+msgid "Registrations on external link"
+msgstr ""
+
+msgid "Pre-registration active"
+msgstr ""
+
+msgid "Registrations opening not set"
+msgstr ""
+
+#, python-format
+msgid "Registrations opening at: %(date)s"
+msgstr ""
+
+msgid "Registrations open"
+msgstr ""
+
+msgid "Filler registrations"
+msgstr ""
+
+msgid "Waiting list registrations"
+msgstr ""
+
+msgid "Registration closed"
+msgstr ""
+
+msgid ""
+"There are no active events, to create a new one access the events management "
+"page:"
+msgstr ""
+
+msgid ""
+"To set up the gateway payment available to players, to let them pay the "
+"registration fee through the platform, access the payment settings "
+"management page:"
+msgstr ""
+
+msgid ""
+"To define which data will be asked in the profile form to the users once "
+"they sign up, access the profile management page:"
+msgstr ""
+
+msgid ""
+"To set the tickets that users can select during registration, access the "
+"tickets management page:"
+msgstr ""
+
+msgid ""
+"To define the registration form, and set up any number of registration "
+"questions and their options, access the registration form management page:"
 msgstr ""
 
 msgid "Reimbursement request item added"
