+++ larpmanager/locale/es/LC_MESSAGES/django.po
@@ -1272,6 +1272,9 @@ msgstr "Escenas previas"
 msgid "surcharge"
 msgstr "recargo"
 
+msgid "this link"
+msgstr "este enlace"
+
 msgid "Accounting"
 msgstr "Contabilidad"
 
@@ -2744,9 +2747,6 @@ msgstr "Entradas adicionales"
 msgid "Approved character"
 msgstr "Carácter aprobado"
 
-msgid "Can't find a user?"
-msgstr "¿No encuentras un usuario?"
-
 msgid "Character portrait"
 msgstr "Retrato del personaje"
 
@@ -3153,9 +3153,6 @@ msgstr "Contabilidad histórica"
 
 msgid "If absent, indicate 0"
 msgstr "Si está ausente, indique 0"
-
-msgid "Links to be accessed:"
-msgstr "Enlaces a los que acceder:"
 
 msgid "List of other users: "
 msgstr "Lista de otros usuarios:"
@@ -4711,6 +4708,9 @@ msgstr "Si está marcada: muestra el eslogan de cada evento"
 
 msgid "If checked: shows the website for each event"
 msgstr "Si está marcada: muestra el sitio web de cada evento"
+
+msgid "If you don't see an user, ask them to access"
+msgstr "Si no ve ningún usuario, pídale que acceda a"
 
 msgid "Manage invoices of payments and approve them"
 msgstr "Gestionar las facturas de los pagos y aprobarlas"
@@ -6343,6 +6343,12 @@ msgid ""
 msgstr "Esta página te permite configurar las funciones activadas"
 
 msgid ""
+"To assign a user to a role, they must first be added to the organization."
+msgstr ""
+"Para asignar un usuario a una función, primero debe añadirse a la "
+"organización."
+
+msgid ""
 "Total amount of payment to be received by this date (0 = all outstanding)"
 msgstr ""
 "Importe total del pago que debe recibirse en esta fecha (0 = todo pendiente)"
@@ -7130,11 +7136,6 @@ msgid ""
 msgstr "Sube tu solicitud firmada y una foto de tu documento de identidad"
 
 msgid ""
-"this figure is the net that should result as the availability of financial "
-"instruments"
-msgstr "esta cifra es el neto que debería resultar del balance financiero"
-
-msgid ""
 "Adds a page displaying member mailing lists, sorted by members' newsletter "
 "preferences."
 msgstr ""
@@ -7685,6 +7686,13 @@ msgid ""
 msgstr ""
 "Opcional: ¡sube una foto tuya asociada a tu personaje específicamente para "
 "este evento!"
+
+msgid ""
+"This is the overall amount that should be held in the organization’s bank "
+"and financial accounts"
+msgstr ""
+"Es la cantidad total que debe mantenerse en las cuentas bancarias y "
+"financieras de la organización."
 
 msgid ""
 "Unique code for internal use - not visible. Indicate a string without spaces"
@@ -8506,13 +8514,6 @@ msgstr ""
 "jugadores que seleccionen ese tipo de entrada"
 
 msgid ""
-"this figure is the associative positive, the operating margin. It takes into"
-" account the budgets of the completed runs"
-msgstr ""
-"esta cifra es el positivo de la organización, su margen operativo. Tiene en "
-"cuenta los presupuestos de los pases realizados"
-
-msgid ""
 "Are you a doctor, a nurse, or a licensed rescuer? We can ask you to "
 "intervene in case accidents occur during the event?"
 msgstr ""
@@ -8635,6 +8636,13 @@ msgid ""
 msgstr ""
 "Se simula la asignación óptima basada en los valores establecidos; puede "
 "ejecutarla tantas veces como desee antes de la carga final."
+
+msgid ""
+"To define which data will be asked in the profile form to the users once "
+"they sign up, access the profile management page:"
+msgstr ""
+"Para definir qué datos se pedirán en el formulario de perfil a los usuarios "
+"una vez que se registren, acceda a la página de gestión de perfiles:"
 
 msgid ""
 "Lets you specify a set of users who receive payment confirmation emails for "
@@ -8788,6 +8796,14 @@ msgid ""
 msgstr ""
 "Esta página muestra las colectas abiertas por los jugadores. Puedes "
 "actualizarlas indicando el destinatario y desembolsar el dinero."
+
+msgid ""
+"This is the associative positive operating margin, taking into account the "
+"budgets of completed runs, minus credits or tokens issued"
+msgstr ""
+"Es el margen de explotación positivo asociativo, teniendo en cuenta los "
+"presupuestos de las ejecuciones realizadas, menos los créditos o fichas "
+"emitidos"
 
 msgid ""
 "Adds a quiz-style system for workshop questions. Organizers can define "
@@ -8973,6 +8989,14 @@ msgstr ""
 "momento."
 
 msgid ""
+"To define the registration form, and set up any number of registration "
+"questions and their options, access the registration form management page:"
+msgstr ""
+"Para definir el formulario de inscripción y configurar cualquier número de "
+"preguntas de inscripción y sus opciones, acceda a la página de gestión del "
+"formulario de inscripción:"
+
+msgid ""
 "If you press once, it will be *included* in the search. If a character "
 "belongs to at least one inclusion category, it will be shown in the search."
 msgstr ""
@@ -9009,14 +9033,6 @@ msgid ""
 msgstr ""
 "Puede realizar el pago en unos minutos <a href= %(url)s'>aquí</a>. Avísanos "
 "si tienes algún problema o si podemos ayudarte en algo."
-
-msgid ""
-"According to LarpManager's privacy rules, you can only view users who have "
-"granted LarpManager permission to share their data with your organisation."
-msgstr ""
-"De acuerdo con las normas de privacidad de LarpManager, sólo puedes ver los "
-"usuarios que han concedido permiso a LarpManager para compartir sus datos "
-"con tu organización."
 
 msgid ""
 "At the top, you'll find shortcuts to access your organization dashboard "
@@ -9102,6 +9118,15 @@ msgstr ""
 "la búsqueda."
 
 msgid ""
+"To set up the gateway payment available to players, to let them pay the "
+"registration fee through the platform, access the payment settings "
+"management page:"
+msgstr ""
+"Para configurar la pasarela de pago disponible para los jugadores, para que "
+"puedan pagar la cuota de inscripción a través de la plataforma, acceda a la "
+"página de gestión de los ajustes de pago:"
+
+msgid ""
 "Indicates how you can be recognized, if you will be found near some point "
 "specific, if you have a lot of luggage: any information that might help us "
 "help you"
@@ -9126,14 +9151,6 @@ msgstr ""
 "Esta página muestra las preguntas enviadas por los usuarios, clasificadas "
 "por los eventos a los que se refieren. Puedes responder a la pregunta o "
 "cerrarla si no requiere realizar ninguna otra acción."
-
-msgid ""
-"Ask your co-workers to access this link; after their confirmation, their "
-"data will be shared with your organisation, and you will be able to add them"
-" to roles."
-msgstr ""
-"Pide a tus compañeros que accedan a este enlace; tras su confirmación, sus "
-"datos se compartirán con tu organización y podrás asignarles cargos."
 
 #, python-format
 msgid ""
