+++ larpmanager/locale/fr/LC_MESSAGES/django.po
@@ -1272,6 +1272,9 @@ msgstr "speedlarp"
 msgid "surcharge"
 msgstr "surtaxe"
 
+msgid "this link"
+msgstr "ce lien"
+
 msgid "Accounting"
 msgstr "Comptabilité"
 
@@ -2744,9 +2747,6 @@ msgstr "Billets supplémentaires"
 msgid "Approved character"
 msgstr "Caractère approuvé"
 
-msgid "Can't find a user?"
-msgstr "Vous ne trouvez pas d'utilisateur ?"
-
 msgid "Character portrait"
 msgstr "Portrait de personnage"
 
@@ -3153,9 +3153,6 @@ msgstr "Comptabilité historique"
 
 msgid "If absent, indicate 0"
 msgstr "En cas d'absence, indiquer 0"
-
-msgid "Links to be accessed:"
-msgstr "Liens à consulter :"
 
 msgid "List of other users: "
 msgstr "Liste des autres utilisateurs :"
@@ -4714,6 +4711,9 @@ msgstr "Si coché : affiche le slogan de chaque événement"
 
 msgid "If checked: shows the website for each event"
 msgstr "Si coché : affiche le site web de chaque événement"
+
+msgid "If you don't see an user, ask them to access"
+msgstr "Si vous ne voyez pas d'utilisateur, demandez-lui d'accéder à"
 
 msgid "Manage invoices of payments and approve them"
 msgstr "Gérer les factures et les paiements et les approuver"
@@ -6430,6 +6430,12 @@ msgid ""
 msgstr "Cette page permet de configurer les fonctionnalités."
 
 msgid ""
+"To assign a user to a role, they must first be added to the organization."
+msgstr ""
+"Pour attribuer un rôle à un utilisateur, celui-ci doit d'abord être ajouté à"
+" l'organisation."
+
+msgid ""
 "Total amount of payment to be received by this date (0 = all outstanding)"
 msgstr "Montant total du paiement à recevoir à cette date (0 = tout est dû)"
 
@@ -7227,13 +7233,6 @@ msgstr ""
 "Téléchargez votre demande signée et une photo de votre pièce d'identité"
 
 msgid ""
-"this figure is the net that should result as the availability of financial "
-"instruments"
-msgstr ""
-"ce chiffre est le montant net qui devrait résulter de la disponibilité des "
-"instruments financiers"
-
-msgid ""
 "Adds a page displaying member mailing lists, sorted by members' newsletter "
 "preferences."
 msgstr ""
@@ -7783,6 +7782,13 @@ msgid ""
 msgstr ""
 "Facultatif : téléchargez une photo de vous associée à votre personnage "
 "spécifiquement pour cet événement !"
+
+msgid ""
+"This is the overall amount that should be held in the organization’s bank "
+"and financial accounts"
+msgstr ""
+"Il s'agit du montant global qui doit être détenu sur les comptes bancaires "
+"et financiers de l'organisation."
 
 msgid ""
 "Unique code for internal use - not visible. Indicate a string without spaces"
@@ -8609,13 +8615,6 @@ msgstr ""
 "qu'aux joueurs qui ont sélectionné ce ticket."
 
 msgid ""
-"this figure is the associative positive, the operating margin. It takes into"
-" account the budgets of the completed runs"
-msgstr ""
-"ce chiffre est le positif associatif, la marge opérationnelle. Il prend en "
-"compte les budgets des séries achevées"
-
-msgid ""
 "Are you a doctor, a nurse, or a licensed rescuer? We can ask you to "
 "intervene in case accidents occur during the event?"
 msgstr ""
@@ -8741,6 +8740,14 @@ msgstr ""
 "L'allocation optimale basée sur les valeurs définies est simulée ; vous "
 "pouvez l'exécuter autant de fois que vous le souhaitez avant le chargement "
 "final."
+
+msgid ""
+"To define which data will be asked in the profile form to the users once "
+"they sign up, access the profile management page:"
+msgstr ""
+"Pour définir les données qui seront demandées aux utilisateurs dans le "
+"formulaire de profil après leur inscription, accédez à la page de gestion "
+"des profils :"
 
 msgid ""
 "Lets you specify a set of users who receive payment confirmation emails for "
@@ -8900,6 +8907,13 @@ msgid ""
 msgstr ""
 "Cette page résume les collectes ouvertes par les joueurs. Vous pouvez les "
 "mettre à jour en indiquant le destinataire et débourser l'argent."
+
+msgid ""
+"This is the associative positive operating margin, taking into account the "
+"budgets of completed runs, minus credits or tokens issued"
+msgstr ""
+"Il s'agit de la marge opérationnelle positive associative, prenant en compte"
+" les budgets des courses réalisées, moins les crédits ou les jetons émis."
 
 msgid ""
 "Adds a quiz-style system for workshop questions. Organizers can define "
@@ -9090,6 +9104,14 @@ msgstr ""
 " pour l'instant !"
 
 msgid ""
+"To define the registration form, and set up any number of registration "
+"questions and their options, access the registration form management page:"
+msgstr ""
+"Pour définir le formulaire d'inscription et paramétrer un nombre quelconque "
+"de questions d'inscription et leurs options, accédez à la page de gestion du"
+" formulaire d'inscription :"
+
+msgid ""
 "If you press once, it will be *included* in the search. If a character "
 "belongs to at least one inclusion category, it will be shown in the search."
 msgstr ""
@@ -9130,14 +9152,6 @@ msgstr ""
 "Vous pouvez effectuer le paiement en quelques minutes <a href= "
 "%(url)s'>ici</a>. N'hésitez pas à nous contacter si vous rencontrez des "
 "problèmes, ou si nous pouvons vous aider d'une manière ou d'une autre !"
-
-msgid ""
-"According to LarpManager's privacy rules, you can only view users who have "
-"granted LarpManager permission to share their data with your organisation."
-msgstr ""
-"Selon les règles de confidentialité de LarpManager, vous ne pouvez voir que "
-"les utilisateurs qui ont autorisé LarpManager à partager leurs données avec "
-"votre organisation."
 
 msgid ""
 "At the top, you'll find shortcuts to access your organization dashboard "
@@ -9225,6 +9239,16 @@ msgstr ""
 " dans la recherche."
 
 msgid ""
+"To set up the gateway payment available to players, to let them pay the "
+"registration fee through the platform, access the payment settings "
+"management page:"
+msgstr ""
+"Pour configurer la passerelle de paiement mise à la disposition des joueurs,"
+" afin de leur permettre de payer les frais d'inscription par l'intermédiaire"
+" de la plate-forme, accédez à la page de gestion des paramètres de paiement "
+":"
+
+msgid ""
 "Indicates how you can be recognized, if you will be found near some point "
 "specific, if you have a lot of luggage: any information that might help us "
 "help you"
@@ -9249,15 +9273,6 @@ msgstr ""
 "Cette page résume les questions soumises par les utilisateurs, réparties "
 "selon les événements auxquels elles se rapportent. Vous pouvez répondre à la"
 " question ou la clôturer si aucune action n'est requise."
-
-msgid ""
-"Ask your co-workers to access this link; after their confirmation, their "
-"data will be shared with your organisation, and you will be able to add them"
-" to roles."
-msgstr ""
-"Demandez à vos collègues d'accéder à ce lien ; après leur confirmation, "
-"leurs données seront partagées avec votre organisation et vous pourrez les "
-"ajouter à des rôles."
 
 #, python-format
 msgid ""
