+++ larpmanager/locale/it/LC_MESSAGES/django.po
@@ -1272,6 +1272,9 @@ msgstr "speedlarp"
 msgid "surcharge"
 msgstr "sovrapprezzo"
 
+msgid "this link"
+msgstr "questo link"
+
 msgid "Accounting"
 msgstr "Contabilità"
 
@@ -2744,9 +2747,6 @@ msgstr "Biglietti aggiuntivi"
 msgid "Approved character"
 msgstr "Personaggio approvato"
 
-msgid "Can't find a user?"
-msgstr "Non riesci a trovare un utente?"
-
 msgid "Character portrait"
 msgstr "Ritratto personaggio"
 
@@ -3153,9 +3153,6 @@ msgstr "Contabilità storica"
 
 msgid "If absent, indicate 0"
 msgstr "Se assente, indicare 0"
-
-msgid "Links to be accessed:"
-msgstr "Link da accedere:"
 
 msgid "List of other users: "
 msgstr "Elenco degli altri utenti:"
@@ -4710,6 +4707,9 @@ msgstr "Se checkato: mostra la tagline per ogni evento"
 
 msgid "If checked: shows the website for each event"
 msgstr "Se checkato: mostra il sito web per ogni evento"
+
+msgid "If you don't see an user, ask them to access"
+msgstr "Se non si vede un utente, chiedergli di accedere a"
 
 msgid "Manage invoices of payments and approve them"
 msgstr "Gestisci le ricevute dei pagamenti e approvale"
@@ -6377,6 +6377,12 @@ msgid ""
 msgstr "Questa pagina ti permette di configurare le funzionalità."
 
 msgid ""
+"To assign a user to a role, they must first be added to the organization."
+msgstr ""
+"Per assegnare un utente a un ruolo, è necessario aggiungerlo "
+"all'organizzazione."
+
+msgid ""
 "Total amount of payment to be received by this date (0 = all outstanding)"
 msgstr ""
 "Ammontare totale di pagamento da ricevere entro questa data (0 = tutto il "
@@ -7175,13 +7181,6 @@ msgstr ""
 "hai indicato"
 
 msgid ""
-"this figure is the net that should result as the availability of financial "
-"instruments"
-msgstr ""
-"questa cifra è il netto che dovrebbe risultare come disponibilità degli "
-"strumenti finanziari"
-
-msgid ""
 "Adds a page displaying member mailing lists, sorted by members' newsletter "
 "preferences."
 msgstr ""
@@ -7731,6 +7730,13 @@ msgid ""
 msgstr ""
 "Opzionale: carica una tua foto associata al tuo personaggio specifica per "
 "questo evento!"
+
+msgid ""
+"This is the overall amount that should be held in the organization’s bank "
+"and financial accounts"
+msgstr ""
+"Si tratta dell'importo complessivo che dovrebbe essere detenuto nei conti "
+"bancari e finanziari dell'organizzazione."
 
 msgid ""
 "Unique code for internal use - not visible. Indicate a string without spaces"
@@ -8551,13 +8557,6 @@ msgstr ""
 "giocatori che selezionano tale biglietto"
 
 msgid ""
-"this figure is the associative positive, the operating margin. It takes into"
-" account the budgets of the completed runs"
-msgstr ""
-"questa cifra è il positivo associativo, il margine operativo. Tiene conto "
-"dei budget delle run in stato concluso"
-
-msgid ""
 "Are you a doctor, a nurse, or a licensed rescuer? We can ask you to "
 "intervene in case accidents occur during the event?"
 msgstr ""
@@ -8679,6 +8678,13 @@ msgid ""
 msgstr ""
 "Viene simulata la assegnazione ottimale in base ai valori impostati; la puoi"
 " eseguire quante volte vuoi prima del caricamento definitivo"
+
+msgid ""
+"To define which data will be asked in the profile form to the users once "
+"they sign up, access the profile management page:"
+msgstr ""
+"Per definire quali dati verranno richiesti nel modulo del profilo agli "
+"utenti una volta registrati, accedere alla pagina di gestione del profilo:"
 
 msgid ""
 "Lets you specify a set of users who receive payment confirmation emails for "
@@ -8833,6 +8839,13 @@ msgid ""
 msgstr ""
 "Questa pagina riepiloga le collette aperte dai giocatori. Puoi aggiornarle "
 "indicando il ricevente ed erogare i soldi versati."
+
+msgid ""
+"This is the associative positive operating margin, taking into account the "
+"budgets of completed runs, minus credits or tokens issued"
+msgstr ""
+"È il margine operativo associativo positivo, che tiene conto dei bilanci "
+"delle corse completate, meno i crediti o i gettoni emessi."
 
 msgid ""
 "Adds a quiz-style system for workshop questions. Organizers can define "
@@ -9018,6 +9031,14 @@ msgstr ""
 "fare nulla in questo momento!"
 
 msgid ""
+"To define the registration form, and set up any number of registration "
+"questions and their options, access the registration form management page:"
+msgstr ""
+"Per definire il modulo di registrazione e impostare un numero qualsiasi di "
+"domande di registrazione e le relative opzioni, accedere alla pagina di "
+"gestione del modulo di registrazione:"
+
+msgid ""
 "If you press once, it will be *included* in the search. If a character "
 "belongs to at least one inclusion category, it will be shown in the search."
 msgstr ""
@@ -9057,14 +9078,6 @@ msgstr ""
 "Puoi eseguire il pagamento in pochissimi minuti <a href='%(url)s'>qui</a>. "
 "Facci sapere se incontri dei problemi, o se possiamo aiutare in qualunque "
 "modo!"
-
-msgid ""
-"According to LarpManager's privacy rules, you can only view users who have "
-"granted LarpManager permission to share their data with your organisation."
-msgstr ""
-"Secondo le regole di privacy di LarpManager, puoi visionare solo gli utenti "
-"che hanno concesso a LarpManager il permesso di condividere i loro dati con "
-"la tua organizzazione."
 
 msgid ""
 "At the top, you'll find shortcuts to access your organization dashboard "
@@ -9152,6 +9165,15 @@ msgstr ""
 "nella ricerca."
 
 msgid ""
+"To set up the gateway payment available to players, to let them pay the "
+"registration fee through the platform, access the payment settings "
+"management page:"
+msgstr ""
+"Per impostare il gateway di pagamento a disposizione dei giocatori, per "
+"consentire loro di pagare la quota di iscrizione attraverso la piattaforma, "
+"accedere alla pagina di gestione delle impostazioni di pagamento:"
+
+msgid ""
 "Indicates how you can be recognized, if you will be found near some point "
 "specific, if you have a lot of luggage: any information that might help us "
 "help you"
@@ -9176,15 +9198,6 @@ msgstr ""
 "Questa pagina riepiloga le domande sottoposte dagli utenti, divise per gli "
 "eventi a cui fanno riferimento. Puoi rispondere alla domanda o chiuderla se "
 "non necessita ulteriori azioni."
-
-msgid ""
-"Ask your co-workers to access this link; after their confirmation, their "
-"data will be shared with your organisation, and you will be able to add them"
-" to roles."
-msgstr ""
-"Chiedi ai tuoi collaboratori di accedere a questo link; dopo la loro "
-"conferma, i loro dati saranno condivisi con la tua organizzazione, e potrai "
-"aggiungerli nei ruoli."
 
 #, python-format
 msgid ""
