+++ larpmanager/locale/nb/LC_MESSAGES/django.po
@@ -1272,6 +1272,9 @@ msgstr "speedlarp"
 msgid "surcharge"
 msgstr "tilleggsavgift"
 
+msgid "this link"
+msgstr "denne lenken"
+
 msgid "Accounting"
 msgstr "Regnskap"
 
@@ -2744,9 +2747,6 @@ msgstr "Ekstra billetter"
 msgid "Approved character"
 msgstr "Godkjent karakter"
 
-msgid "Can't find a user?"
-msgstr "Finner du ikke en bruker?"
-
 msgid "Character portrait"
 msgstr "Portrett av en person"
 
@@ -3153,9 +3153,6 @@ msgstr "Historisk regnskap"
 
 msgid "If absent, indicate 0"
 msgstr "Hvis fraværende, angi 0"
-
-msgid "Links to be accessed:"
-msgstr "Lenker som skal åpnes:"
 
 msgid "List of other users: "
 msgstr "Liste over andre brukere:"
@@ -4712,6 +4709,9 @@ msgstr "Hvis avkrysset: viser tagline for hvert arrangement"
 
 msgid "If checked: shows the website for each event"
 msgstr "Hvis avkrysset: viser nettstedet for hvert arrangement"
+
+msgid "If you don't see an user, ask them to access"
+msgstr "Hvis du ikke ser en bruker, kan du be dem om å få tilgang til"
 
 msgid "Manage invoices of payments and approve them"
 msgstr "Administrere og godkjenne fakturaer og betalinger"
@@ -6298,6 +6298,12 @@ msgid ""
 msgstr "På denne siden kan du konfigurere funksjoner."
 
 msgid ""
+"To assign a user to a role, they must first be added to the organization."
+msgstr ""
+"For å tilordne en bruker til en rolle må vedkommende først legges til i "
+"organisasjonen."
+
+msgid ""
 "Total amount of payment to be received by this date (0 = all outstanding)"
 msgstr "Totalt beløp som skal mottas innen denne datoen (0 = alt utestående)"
 
@@ -7049,13 +7055,6 @@ msgstr ""
 "ditt."
 
 msgid ""
-"this figure is the net that should result as the availability of financial "
-"instruments"
-msgstr ""
-"dette tallet er det nettoresultatet som bør fremkomme som følge av "
-"tilgjengeligheten av finansielle instrumenter"
-
-msgid ""
 "Adds a page displaying member mailing lists, sorted by members' newsletter "
 "preferences."
 msgstr ""
@@ -7603,6 +7602,13 @@ msgid ""
 msgstr ""
 "Valgfritt: Last opp et bilde av deg selv i forbindelse med karakteren din "
 "spesielt for dette arrangementet!"
+
+msgid ""
+"This is the overall amount that should be held in the organization’s bank "
+"and financial accounts"
+msgstr ""
+"Dette er det totale beløpet som skal stå på organisasjonens bank- og "
+"finanskontoer"
 
 msgid ""
 "Unique code for internal use - not visible. Indicate a string without spaces"
@@ -8421,13 +8427,6 @@ msgstr ""
 "spillere som har valgt denne billetten."
 
 msgid ""
-"this figure is the associative positive, the operating margin. It takes into"
-" account the budgets of the completed runs"
-msgstr ""
-"dette tallet er det assosiative positive, driftsmarginen. Det tar hensyn til"
-" budsjettene for de fullførte kjøringene"
-
-msgid ""
 "Are you a doctor, a nurse, or a licensed rescuer? We can ask you to "
 "intervene in case accidents occur during the event?"
 msgstr ""
@@ -8549,6 +8548,13 @@ msgid ""
 msgstr ""
 "Den optimale allokeringen basert på de innstilte verdiene simuleres; du kan "
 "kjøre den så mange ganger du vil før den endelige innlastingen"
+
+msgid ""
+"To define which data will be asked in the profile form to the users once "
+"they sign up, access the profile management page:"
+msgstr ""
+"For å definere hvilke data som skal spørres om i profilskjemaet til brukerne"
+" når de registrerer seg, går du til siden for profiladministrasjon:"
 
 msgid ""
 "Lets you specify a set of users who receive payment confirmation emails for "
@@ -8702,6 +8708,13 @@ msgid ""
 msgstr ""
 "Denne siden oppsummerer innsamlingene som er åpnet av spillere. Du kan "
 "oppdatere dem ved å angi mottaker og utbetale pengene."
+
+msgid ""
+"This is the associative positive operating margin, taking into account the "
+"budgets of completed runs, minus credits or tokens issued"
+msgstr ""
+"Dette er den assosiative positive driftsmarginen, som tar hensyn til "
+"budsjettene for fullførte kjøringer, minus utstedte kreditter eller tokens"
 
 msgid ""
 "Adds a quiz-style system for workshop questions. Organizers can define "
@@ -8883,6 +8896,14 @@ msgstr ""
 "tidspunktet!"
 
 msgid ""
+"To define the registration form, and set up any number of registration "
+"questions and their options, access the registration form management page:"
+msgstr ""
+"Du kan definere registreringsskjemaet og sette opp et hvilket som helst "
+"antall registreringsspørsmål og alternativer for disse ved å gå til siden "
+"for administrasjon av registreringsskjema:"
+
+msgid ""
 "If you press once, it will be *included* in the search. If a character "
 "belongs to at least one inclusion category, it will be shown in the search."
 msgstr ""
@@ -8920,13 +8941,6 @@ msgstr ""
 "Du kan foreta betalingen på bare noen få minutter <a href= %(url)s'>her</a>."
 " Gi oss beskjed hvis du støter på problemer, eller hvis vi kan hjelpe deg på"
 " noen måte!"
-
-msgid ""
-"According to LarpManager's privacy rules, you can only view users who have "
-"granted LarpManager permission to share their data with your organisation."
-msgstr ""
-"I henhold til LarpManagers personvernregler kan du bare se brukere som har "
-"gitt LarpManager tillatelse til å dele dataene deres med organisasjonen din."
 
 msgid ""
 "At the top, you'll find shortcuts to access your organization dashboard "
@@ -9009,6 +9023,15 @@ msgstr ""
 "tilhører én ekskluderingskategori, vises det ikke i søket."
 
 msgid ""
+"To set up the gateway payment available to players, to let them pay the "
+"registration fee through the platform, access the payment settings "
+"management page:"
+msgstr ""
+"For å konfigurere gateway-betalingen som er tilgjengelig for spillere, slik "
+"at de kan betale registreringsavgiften via plattformen, åpner du siden for "
+"administrering av betalingsinnstillinger:"
+
+msgid ""
 "Indicates how you can be recognized, if you will be found near some point "
 "specific, if you have a lot of luggage: any information that might help us "
 "help you"
@@ -9033,15 +9056,6 @@ msgstr ""
 "Denne siden oppsummerer spørsmålene som brukerne har sendt inn, inndelt "
 "etter hendelsene de gjelder. Du kan svare på spørsmålet eller lukke det hvis"
 " det ikke er behov for ytterligere tiltak."
-
-msgid ""
-"Ask your co-workers to access this link; after their confirmation, their "
-"data will be shared with your organisation, and you will be able to add them"
-" to roles."
-msgstr ""
-"Be medarbeiderne dine om å få tilgang til denne lenken. Når de har bekreftet"
-" dette, vil dataene deres bli delt med organisasjonen din, og du vil kunne "
-"legge dem til i roller."
 
 #, python-format
 msgid ""
