+++ larpmanager/locale/nl/LC_MESSAGES/django.po
@@ -1272,6 +1272,9 @@ msgstr "speedlarp"
 msgid "surcharge"
 msgstr "toeslag"
 
+msgid "this link"
+msgstr "deze link"
+
 msgid "Accounting"
 msgstr "Boekhouding"
 
@@ -2744,9 +2747,6 @@ msgstr "Extra tickets"
 msgid "Approved character"
 msgstr "Goedgekeurd karakter"
 
-msgid "Can't find a user?"
-msgstr "Kun je geen gebruiker vinden?"
-
 msgid "Character portrait"
 msgstr "Karakterportret"
 
@@ -3153,9 +3153,6 @@ msgstr "Historische boekhouding"
 
 msgid "If absent, indicate 0"
 msgstr "Indien afwezig, geef 0 aan"
-
-msgid "Links to be accessed:"
-msgstr "Te openen links:"
 
 msgid "List of other users: "
 msgstr "Lijst van andere gebruikers:"
@@ -4714,6 +4711,9 @@ msgstr "Indien aangevinkt: toont de tagline voor elk evenement"
 
 msgid "If checked: shows the website for each event"
 msgstr "Indien aangevinkt: toont de website voor elk evenement"
+
+msgid "If you don't see an user, ask them to access"
+msgstr "Als u geen gebruiker ziet, vraag hem dan om toegang te krijgen tot"
 
 msgid "Manage invoices of payments and approve them"
 msgstr "Facturen van betalingen beheren en goedkeuren"
@@ -6395,6 +6395,12 @@ msgid ""
 msgstr "Op deze pagina kunt u functies configureren."
 
 msgid ""
+"To assign a user to a role, they must first be added to the organization."
+msgstr ""
+"Om een gebruiker aan een rol toe te wijzen, moet hij eerst worden toegevoegd"
+" aan de organisatie."
+
+msgid ""
 "Total amount of payment to be received by this date (0 = all outstanding)"
 msgstr ""
 "Totaalbedrag van de betaling te ontvangen tegen deze datum (0 = alles "
@@ -7187,13 +7193,6 @@ msgid ""
 msgstr "Upload uw ondertekende aanvraag en een foto van uw identiteitsbewijs"
 
 msgid ""
-"this figure is the net that should result as the availability of financial "
-"instruments"
-msgstr ""
-"dit cijfer is het nettoresultaat dat zou moeten resulteren uit de "
-"beschikbaarheid van financiële instrumenten"
-
-msgid ""
 "Adds a page displaying member mailing lists, sorted by members' newsletter "
 "preferences."
 msgstr ""
@@ -7745,6 +7744,13 @@ msgid ""
 msgstr ""
 "Optioneel: upload een foto van jezelf in combinatie met je personage "
 "speciaal voor dit evenement!"
+
+msgid ""
+"This is the overall amount that should be held in the organization’s bank "
+"and financial accounts"
+msgstr ""
+"Dit is het totale bedrag dat op de bank- en financiële rekeningen van de "
+"organisatie moet staan"
 
 msgid ""
 "Unique code for internal use - not visible. Indicate a string without spaces"
@@ -8568,13 +8574,6 @@ msgstr ""
 "spelers die dat ticket hebben geselecteerd."
 
 msgid ""
-"this figure is the associative positive, the operating margin. It takes into"
-" account the budgets of the completed runs"
-msgstr ""
-"Dit cijfer is de positieve associatie, de operationele marge. Het houdt "
-"rekening met de budgetten van de voltooide runs."
-
-msgid ""
 "Are you a doctor, a nurse, or a licensed rescuer? We can ask you to "
 "intervene in case accidents occur during the event?"
 msgstr ""
@@ -8699,6 +8698,13 @@ msgstr ""
 "De optimale toewijzing op basis van de ingestelde waarden wordt gesimuleerd;"
 " je kunt het zo vaak uitvoeren als je wilt voordat het definitief wordt "
 "geladen."
+
+msgid ""
+"To define which data will be asked in the profile form to the users once "
+"they sign up, access the profile management page:"
+msgstr ""
+"Om te definiëren welke gegevens worden gevraagd in het profielformulier aan "
+"de gebruikers zodra ze zich aanmelden, gaat u naar de pagina Profielbeheer:"
 
 msgid ""
 "Lets you specify a set of users who receive payment confirmation emails for "
@@ -8854,6 +8860,13 @@ msgstr ""
 "Deze pagina geeft een overzicht van de collecties die geopend zijn door "
 "spelers. Je kunt ze bijwerken door de ontvanger aan te geven en het geld uit"
 " te betalen."
+
+msgid ""
+"This is the associative positive operating margin, taking into account the "
+"budgets of completed runs, minus credits or tokens issued"
+msgstr ""
+"Dit is de associatieve positieve operationele marge, rekening houdend met de"
+" budgetten van voltooide runs, minus uitgegeven credits of tokens."
 
 msgid ""
 "Adds a quiz-style system for workshop questions. Organizers can define "
@@ -9041,6 +9054,14 @@ msgstr ""
 "moment niets te doen!"
 
 msgid ""
+"To define the registration form, and set up any number of registration "
+"questions and their options, access the registration form management page:"
+msgstr ""
+"Ga naar de beheerpagina voor registratieformulieren om het "
+"registratieformulier te definiëren en een willekeurig aantal "
+"registratievragen en hun opties in te stellen:"
+
+msgid ""
 "If you press once, it will be *included* in the search. If a character "
 "belongs to at least one inclusion category, it will be shown in the search."
 msgstr ""
@@ -9081,14 +9102,6 @@ msgstr ""
 "U kunt de betaling in slechts een paar minuten <a href= %(url)s'>hier</a>. "
 "Laat het ons weten als je problemen ondervindt of als we je op een of andere"
 " manier kunnen helpen!"
-
-msgid ""
-"According to LarpManager's privacy rules, you can only view users who have "
-"granted LarpManager permission to share their data with your organisation."
-msgstr ""
-"Volgens de privacyregels van LarpManager kun je alleen gebruikers bekijken "
-"die LarpManager toestemming hebben gegeven om hun gegevens met jouw "
-"organisatie te delen."
 
 msgid ""
 "At the top, you'll find shortcuts to access your organization dashboard "
@@ -9174,6 +9187,15 @@ msgstr ""
 "niet getoond in de zoekopdracht."
 
 msgid ""
+"To set up the gateway payment available to players, to let them pay the "
+"registration fee through the platform, access the payment settings "
+"management page:"
+msgstr ""
+"Om de gatewaybetaling in te stellen die beschikbaar is voor spelers, zodat "
+"ze het inschrijfgeld via het platform kunnen betalen, ga je naar de "
+"beheerpagina met betalingsinstellingen:"
+
+msgid ""
 "Indicates how you can be recognized, if you will be found near some point "
 "specific, if you have a lot of luggage: any information that might help us "
 "help you"
@@ -9199,15 +9221,6 @@ msgstr ""
 "ingediend, onderverdeeld naar de gebeurtenissen waar ze betrekking op "
 "hebben. Je kunt de vraag beantwoorden of sluiten als er geen verdere actie "
 "nodig is."
-
-msgid ""
-"Ask your co-workers to access this link; after their confirmation, their "
-"data will be shared with your organisation, and you will be able to add them"
-" to roles."
-msgstr ""
-"Vraag uw collega's om toegang te krijgen tot deze link; na hun bevestiging "
-"worden hun gegevens gedeeld met uw organisatie en kunt u ze toevoegen aan "
-"rollen."
 
 #, python-format
 msgid ""
