+++ larpmanager/locale/pl/LC_MESSAGES/django.po
@@ -1272,6 +1272,9 @@ msgstr "speedlarp"
 msgid "surcharge"
 msgstr "dopłata"
 
+msgid "this link"
+msgstr "ten link"
+
 msgid "Accounting"
 msgstr "Księgowość"
 
@@ -2744,9 +2747,6 @@ msgstr "Dodatkowe bilety"
 msgid "Approved character"
 msgstr "Zatwierdzony charakter"
 
-msgid "Can't find a user?"
-msgstr "Nie możesz znaleźć użytkownika?"
-
 msgid "Character portrait"
 msgstr "Portret postaci"
 
@@ -3153,9 +3153,6 @@ msgstr "Księgowość historyczna"
 
 msgid "If absent, indicate 0"
 msgstr "Jeśli nieobecny, wskaż 0"
-
-msgid "Links to be accessed:"
-msgstr "Dostępne linki:"
 
 msgid "List of other users: "
 msgstr "Lista innych użytkowników:"
@@ -4710,6 +4707,9 @@ msgstr "Jeśli zaznaczone: wyświetla slogan dla każdego wydarzenia."
 
 msgid "If checked: shows the website for each event"
 msgstr "Jeśli zaznaczone: wyświetla stronę internetową dla każdego wydarzenia"
+
+msgid "If you don't see an user, ask them to access"
+msgstr "Jeśli nie widzisz użytkownika, poproś go o dostęp"
 
 msgid "Manage invoices of payments and approve them"
 msgstr "Zarządzanie fakturami i płatnościami oraz ich zatwierdzanie"
@@ -6285,6 +6285,12 @@ msgid ""
 msgstr "Ta strona umożliwia edycję konfiguracji aktywowanych funkcji."
 
 msgid ""
+"To assign a user to a role, they must first be added to the organization."
+msgstr ""
+"Aby przypisać użytkownika do roli, musi on najpierw zostać dodany do "
+"organizacji."
+
+msgid ""
 "Total amount of payment to be received by this date (0 = all outstanding)"
 msgstr ""
 "Całkowita kwota płatności do otrzymania do tego dnia (0 = wszystkie zaległe)"
@@ -7061,13 +7067,6 @@ msgstr ""
 "dokumentu PDF)."
 
 msgid ""
-"this figure is the net that should result as the availability of financial "
-"instruments"
-msgstr ""
-"Liczba ta jest wartością netto, która powinna wynikać z dostępności "
-"instrumentów finansowych"
-
-msgid ""
 "Adds a page displaying member mailing lists, sorted by members' newsletter "
 "preferences."
 msgstr ""
@@ -7617,6 +7616,13 @@ msgid ""
 msgstr ""
 "Opcjonalnie: prześlij swoje zdjęcie związane z postacią specjalnie na "
 "potrzeby tego wydarzenia!"
+
+msgid ""
+"This is the overall amount that should be held in the organization’s bank "
+"and financial accounts"
+msgstr ""
+"Jest to całkowita kwota, która powinna znajdować się na rachunkach bankowych"
+" i finansowych organizacji."
 
 msgid ""
 "Unique code for internal use - not visible. Indicate a string without spaces"
@@ -8435,13 +8441,6 @@ msgstr ""
 "tylko graczom, którzy wybrali jeden z tych biletów"
 
 msgid ""
-"this figure is the associative positive, the operating margin. It takes into"
-" account the budgets of the completed runs"
-msgstr ""
-"Liczba ta jest dodatnią wartością asocjacyjną, marżą operacyjną. Uwzględnia "
-"ona budżety ukończonych serii"
-
-msgid ""
 "Are you a doctor, a nurse, or a licensed rescuer? We can ask you to "
 "intervene in case accidents occur during the event?"
 msgstr ""
@@ -8563,6 +8562,13 @@ msgid ""
 msgstr ""
 "Symulowana jest optymalna alokacja oparta na ustawionych wartościach; można "
 "ją uruchomić dowolną liczbę razy przed ostatecznym załadowaniem"
+
+msgid ""
+"To define which data will be asked in the profile form to the users once "
+"they sign up, access the profile management page:"
+msgstr ""
+"Aby zdefiniować dane, o które użytkownicy będą pytani w formularzu profilu "
+"po zarejestrowaniu się, przejdź do strony zarządzania profilem:"
 
 msgid ""
 "Lets you specify a set of users who receive payment confirmation emails for "
@@ -8713,6 +8719,13 @@ msgid ""
 msgstr ""
 "Ta strona pokazuje zbiórki otwarte przez graczy. Możesz je aktualizować, "
 "wskazując odbiorcę i wypłacając wpłacone pieniądze."
+
+msgid ""
+"This is the associative positive operating margin, taking into account the "
+"budgets of completed runs, minus credits or tokens issued"
+msgstr ""
+"Jest to skojarzona dodatnia marża operacyjna, uwzględniająca budżety "
+"ukończonych serii, pomniejszona o wydane kredyty lub tokeny"
 
 msgid ""
 "Adds a quiz-style system for workshop questions. Organizers can define "
@@ -8897,6 +8910,14 @@ msgstr ""
 "sprawdzane przez personel. W tym momencie nie musisz nic robić!"
 
 msgid ""
+"To define the registration form, and set up any number of registration "
+"questions and their options, access the registration form management page:"
+msgstr ""
+"Aby zdefiniować formularz rejestracyjny i skonfigurować dowolną liczbę pytań"
+" rejestracyjnych i ich opcji, przejdź do strony zarządzania formularzem "
+"rejestracyjnym:"
+
+msgid ""
 "If you press once, it will be *included* in the search. If a character "
 "belongs to at least one inclusion category, it will be shown in the search."
 msgstr ""
@@ -8937,14 +8958,6 @@ msgstr ""
 "Możesz dokonać płatności w ciągu kilku minut <a href= %(url)s\">tutaj</a>. "
 "Daj nam znać, jeśli napotkasz jakiekolwiek problemy lub jeśli możemy w "
 "jakikolwiek sposób pomóc!"
-
-msgid ""
-"According to LarpManager's privacy rules, you can only view users who have "
-"granted LarpManager permission to share their data with your organisation."
-msgstr ""
-"Zgodnie z zasadami prywatności LarpManager, możesz przeglądać tylko tych "
-"użytkowników, którzy udzielili LarpManager pozwolenia na udostępnianie "
-"swoich danych Twojej organizacji."
 
 msgid ""
 "At the top, you'll find shortcuts to access your organization dashboard "
@@ -9029,6 +9042,15 @@ msgstr ""
 "zostanie wyświetlony w wyszukiwaniu."
 
 msgid ""
+"To set up the gateway payment available to players, to let them pay the "
+"registration fee through the platform, access the payment settings "
+"management page:"
+msgstr ""
+"Aby skonfigurować bramkę płatności dostępną dla graczy, aby umożliwić im "
+"uiszczenie opłaty rejestracyjnej za pośrednictwem platformy, przejdź do "
+"strony zarządzania ustawieniami płatności:"
+
+msgid ""
 "Indicates how you can be recognized, if you will be found near some point "
 "specific, if you have a lot of luggage: any information that might help us "
 "help you"
@@ -9053,15 +9075,6 @@ msgstr ""
 "Ta strona pokazuje pytania przesłane przez użytkowników, podzielone według "
 "wydarzeń, do których się odnoszą. Możesz odpowiedzieć na pytanie lub zamknąć"
 " je, jeśli nie wymaga dalszych działań."
-
-msgid ""
-"Ask your co-workers to access this link; after their confirmation, their "
-"data will be shared with your organisation, and you will be able to add them"
-" to roles."
-msgstr ""
-"Poproś swoich współpracowników o dostęp do tego linku; po potwierdzeniu ich "
-"dane zostaną udostępnione Twojej organizacji, a Ty będziesz mógł dodawać ich"
-" do ról."
 
 #, python-format
 msgid ""
