+++ larpmanager/locale/sv/LC_MESSAGES/django.po
@@ -1272,6 +1272,9 @@ msgstr "Speedlarp"
 msgid "surcharge"
 msgstr "tilläggsavgift"
 
+msgid "this link"
+msgstr "denna länk"
+
 msgid "Accounting"
 msgstr "Redovisning"
 
@@ -2744,9 +2747,6 @@ msgstr "Ytterligare biljetter"
 msgid "Approved character"
 msgstr "Godkänd karaktär"
 
-msgid "Can't find a user?"
-msgstr "Hittar du inte en användare?"
-
 msgid "Character portrait"
 msgstr "Porträtt av karaktären"
 
@@ -3153,9 +3153,6 @@ msgstr "Historisk redovisning"
 
 msgid "If absent, indicate 0"
 msgstr "Om frånvarande, ange 0"
-
-msgid "Links to be accessed:"
-msgstr "Länkar som ska nås:"
 
 msgid "List of other users: "
 msgstr "Lista över andra användare:"
@@ -4712,6 +4709,9 @@ msgstr "Om markerat: visar tagline för varje evenemang"
 
 msgid "If checked: shows the website for each event"
 msgstr "Om markerat: visar webbplatsen för varje evenemang"
+
+msgid "If you don't see an user, ask them to access"
+msgstr "Om du inte ser någon användare, be dem att gå in på"
 
 msgid "Manage invoices of payments and approve them"
 msgstr "Hantera fakturor och betalningar och godkänna dem"
@@ -6293,6 +6293,12 @@ msgid ""
 msgstr "På den här sidan kan du konfigurera funktioner."
 
 msgid ""
+"To assign a user to a role, they must first be added to the organization."
+msgstr ""
+"För att tilldela en användare en roll måste denne först läggas till i "
+"organisationen."
+
+msgid ""
 "Total amount of payment to be received by this date (0 = all outstanding)"
 msgstr ""
 "Totalt belopp som ska betalas senast detta datum (0 = allt utestående)"
@@ -7071,13 +7077,6 @@ msgstr ""
 "Ladda upp din undertecknade ansökan och ett foto av din identitetshandling"
 
 msgid ""
-"this figure is the net that should result as the availability of financial "
-"instruments"
-msgstr ""
-"denna siffra är det netto som bör bli resultatet av tillgången till "
-"finansiella instrument"
-
-msgid ""
 "Adds a page displaying member mailing lists, sorted by members' newsletter "
 "preferences."
 msgstr ""
@@ -7624,6 +7623,13 @@ msgid ""
 msgstr ""
 "Valfritt: ladda upp ett foto av dig själv tillsammans med din rollfigur "
 "specifikt för detta evenemang!"
+
+msgid ""
+"This is the overall amount that should be held in the organization’s bank "
+"and financial accounts"
+msgstr ""
+"Detta är det totala belopp som ska finnas på organisationens bank- och "
+"finanskonton"
 
 msgid ""
 "Unique code for internal use - not visible. Indicate a string without spaces"
@@ -8442,13 +8448,6 @@ msgstr ""
 "spelare som väljer den biljetten"
 
 msgid ""
-"this figure is the associative positive, the operating margin. It takes into"
-" account the budgets of the completed runs"
-msgstr ""
-"Denna siffra är det positiva associativa resultatet, rörelsemarginalen. Den "
-"tar hänsyn till budgetarna för de avslutade körningarna"
-
-msgid ""
 "Are you a doctor, a nurse, or a licensed rescuer? We can ask you to "
 "intervene in case accidents occur during the event?"
 msgstr ""
@@ -8570,6 +8569,13 @@ msgid ""
 msgstr ""
 "Den optimala fördelningen baserat på de inställda värdena simuleras; du kan "
 "köra den så många gånger du vill innan den slutliga laddningen"
+
+msgid ""
+"To define which data will be asked in the profile form to the users once "
+"they sign up, access the profile management page:"
+msgstr ""
+"För att definiera vilka uppgifter som ska efterfrågas i profilformuläret för"
+" användarna när de registrerar sig, gå till sidan för profilhantering:"
 
 msgid ""
 "Lets you specify a set of users who receive payment confirmation emails for "
@@ -8722,6 +8728,14 @@ msgid ""
 msgstr ""
 "Denna sida sammanfattar de samlingar som öppnats av spelare. Du kan "
 "uppdatera dem genom att ange mottagaren och betala ut pengarna."
+
+msgid ""
+"This is the associative positive operating margin, taking into account the "
+"budgets of completed runs, minus credits or tokens issued"
+msgstr ""
+"Detta är den associativa positiva rörelsemarginalen, med hänsyn tagen till "
+"budgetarna för genomförda körningar, minus utfärdade krediter eller "
+"polletter"
 
 msgid ""
 "Adds a quiz-style system for workshop questions. Organizers can define "
@@ -8902,6 +8916,14 @@ msgstr ""
 " och den granskas av personalen. Du behöver inte göra någonting just nu!"
 
 msgid ""
+"To define the registration form, and set up any number of registration "
+"questions and their options, access the registration form management page:"
+msgstr ""
+"Om du vill definiera registreringsformuläret och ställa in valfritt antal "
+"registreringsfrågor och deras alternativ går du till sidan för hantering av "
+"registreringsformulär:"
+
+msgid ""
 "If you press once, it will be *included* in the search. If a character "
 "belongs to at least one inclusion category, it will be shown in the search."
 msgstr ""
@@ -8939,13 +8961,6 @@ msgstr ""
 "Du kan göra betalningen på bara några minuter <a href= %(url)s'>här</a>. Låt"
 " oss veta om du stöter på några problem eller om vi kan hjälpa till på något"
 " sätt!"
-
-msgid ""
-"According to LarpManager's privacy rules, you can only view users who have "
-"granted LarpManager permission to share their data with your organisation."
-msgstr ""
-"Enligt LarpManagers sekretessregler kan du bara se användare som har gett "
-"LarpManager tillstånd att dela sina uppgifter med din organisation."
 
 msgid ""
 "At the top, you'll find shortcuts to access your organization dashboard "
@@ -9029,6 +9044,15 @@ msgstr ""
 "sökningen."
 
 msgid ""
+"To set up the gateway payment available to players, to let them pay the "
+"registration fee through the platform, access the payment settings "
+"management page:"
+msgstr ""
+"För att ställa in den gateway-betalning som är tillgänglig för spelare, så "
+"att de kan betala registreringsavgiften via plattformen, öppnar du "
+"hanteringssidan för betalningsinställningar:"
+
+msgid ""
 "Indicates how you can be recognized, if you will be found near some point "
 "specific, if you have a lot of luggage: any information that might help us "
 "help you"
@@ -9052,15 +9076,6 @@ msgstr ""
 "På denna sida sammanfattas de frågor som användarna har ställt, indelade "
 "efter de händelser som de avser. Du kan besvara frågan eller stänga den om "
 "ingen ytterligare åtgärd krävs."
-
-msgid ""
-"Ask your co-workers to access this link; after their confirmation, their "
-"data will be shared with your organisation, and you will be able to add them"
-" to roles."
-msgstr ""
-"Be dina medarbetare att gå in på den här länken. När de har bekräftat detta "
-"kommer deras uppgifter att delas med din organisation och du kan lägga till "
-"dem i roller."
 
 #, python-format
 msgid ""
