+++ larpmanager/templates/larpmanager/manage/orga.html
@@ -36,7 +36,7 @@
             <h3>
                 <a href="{% url 'orga_accounting' run.event.slug run.number %}">{% trans "Accounting" %}</a>
             </h3>
-            <table class="mob">
+            <table>
                 <tr>
                     <th descr="{% trans "Total amount of money received for this event, minus commission fees and refunds" %}">
                         <a href="#" class="sidebar-link">{% trans "Total revenue" %}</a>
