+++ larpmanager/urls/user.py
@@ -18,17 +18,9 @@
 #
 # SPDX-License-Identifier: AGPL-3.0-or-later OR Proprietary
 
-from django.conf import (
-    settings as conf_settings,
-)
-from django.contrib.auth import (
-    views as auth_views,
-)
-from django.urls import (
-    include,
-    path,
-    re_path,
-)
+from django.conf import settings as conf_settings
+from django.contrib.auth import views as auth_views
+from django.urls import include, path, re_path
 
 from larpmanager.forms.member import (
     MyAuthForm,
@@ -38,30 +30,14 @@ from larpmanager.forms.member import (
 )
 from larpmanager.views import auth as views_auth
 from larpmanager.views import base as views_base
-from larpmanager.views.user import (
-    accounting as views_ua,
-)
-from larpmanager.views.user import (
-    casting as views_uca,
-)
-from larpmanager.views.user import (
-    character as views_uc,
-)
-from larpmanager.views.user import (
-    event as views_ue,
-)
-from larpmanager.views.user import (
-    member as views_um,
-)
-from larpmanager.views.user import (
-    miscellanea as views_ums,
-)
-from larpmanager.views.user import (
-    pdf as views_up,
-)
-from larpmanager.views.user import (
-    registration as views_ur,
-)
+from larpmanager.views.user import accounting as views_ua
+from larpmanager.views.user import casting as views_uca
+from larpmanager.views.user import character as views_uc
+from larpmanager.views.user import event as views_ue
+from larpmanager.views.user import member as views_um
+from larpmanager.views.user import miscellanea as views_ums
+from larpmanager.views.user import pdf as views_up
+from larpmanager.views.user import registration as views_ur
 
 urlpatterns = [
     path(
@@ -392,16 +368,6 @@ urlpatterns = [
     re_path(
         r"^paypal/",
         include("paypal.standard.ipn.urls"),
-    ),
-    path(
-        "manage/",
-        views_base.manage,
-        name="manage",
-    ),
-    path(
-        "<slug:s>/<int:n>/manage/",
-        views_base.manage,
-        name="manage",
     ),
     path(
         "<slug:s>/<int:n>/",
